+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project somewhat adheres to [Semantic Versioning](https://semver.org/sp
 The **"Breaking Changes"** listed below are changes that have been made in the decompilation projects (e.g. pokeemerald), which porymap requires in order to work properly. It also includes changes to the scripting API that may change the behavior of existing porymap scripts. If porymap is used with a project or API script that is not up-to-date with the breaking changes, then porymap will likely break or behave improperly.
 
 ## [Unreleased]
-Nothing, yet.
+### Fixed
+- Fix `Add Region Map...` not updating the region map settings file.
+- Fix some crashes on invalid region map tilesets.
+- Improve error reporting for invalid region map editor settings.
 
 ## [5.4.1] - 2024-03-21
 ### Fixed